+++ packages/polaris-viz/src/components/TrendIndicator/index.ts
@@ -1,2 +1,4 @@
 export {TrendIndicator} from './TrendIndicator';
 export type {TrendIndicatorProps} from './TrendIndicator';
+export {estimateTrendIndicatorWidth} from './utilities/estimateTrendIndicatorWidth';
+export {HEIGHT as TREND_INDICATOR_HEIGHT} from './constants';